On fullfiled and on rejected handler should execute asynchronously
Previously, handlers were executed immediately after resolving or
rejecting a promise. So, you would have different behaviour depending
on when you added a handler. This means that we were not fully compliant
with point 2.2.4 of the Promises A+ spec.

From now on, a promise takes a Task Scheduler object that should implement a
`#schedule_execution` method. This method should take care of providing
the execution context for handlers.

We've also removed readers for value and reason, as we should encourage
using handlers when trying to read the value of a promise.